cloudFunction/cargaIncrementalWeather/requirements.txt
google-cloud-bigquery>=3.11.3
google-cloud-storage>=2.10.0
pandas>=2.0.3
python-dateutil>=2.8.2